requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pandas==2.3.2
passlib==1.7.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
stripe==5.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
# Email and Communication
sendgrid==6.10.0
twilio==8.5.0
# Template Engine (already included via Flask)
# File Processing
python-magic==0.4.27
Pillow==10.0.0
# Security
cryptography==41.0.4

# AI/ML and NLP for Chatbot
transformers==4.36.0
torch==2.1.1
torchvision==0.16.1
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
sentence-transformers==2.2.2
langdetect==1.0.9
textblob==0.17.1

# Real-time Communication
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3

# Redis for caching and session management
redis==5.0.1

# Additional AI utilities
openai==1.3.7
huggingface-hub==0.19.4
